Move methods from LocalNode to Client.
