+++ linera-core/src/client/mod.rs
@@ -56,6 +56,7 @@ use linera_execution::{
 };
 use linera_storage::{Clock as _, Storage};
 use linera_views::views::ViewError;
+use rand::prelude::SliceRandom as _;
 use serde::Serialize;
 use thiserror::Error;
 use tokio::sync::OwnedRwLockReadGuard;
@@ -292,26 +293,119 @@ where
     P: ValidatorNodeProvider + Sync + 'static,
     S: Storage + Sync + Send + Clone + 'static,
 {
-    async fn download_certificates(
+    /// Downloads and processes all certificates up to (excluding) the specified height.
+    #[instrument(level = "trace", skip(self, validators))]
+    pub async fn download_certificates(
         &self,
-        nodes: &[RemoteNode<P::Node>],
+        validators: &[RemoteNode<impl ValidatorNode>],
         chain_id: ChainId,
-        height: BlockHeight,
-    ) -> Result<Box<ChainInfo>, LocalNodeError> {
-        self.local_node
-            .download_certificates(nodes, chain_id, height, &self.notifier)
-            .await
+        target_next_block_height: BlockHeight,
+    ) -> Result<Box<ChainInfo>, ChainClientError> {
+        // Sequentially try each validator in random order.
+        let mut validators = validators.iter().collect::<Vec<_>>();
+        validators.shuffle(&mut rand::thread_rng());
+        for remote_node in validators {
+            let info = self.local_node.local_chain_info(chain_id).await?;
+            if target_next_block_height <= info.next_block_height {
+                return Ok(info);
+            }
+            self.try_download_certificates_from(
+                remote_node,
+                chain_id,
+                info.next_block_height,
+                target_next_block_height,
+            )
+            .await?;
+        }
+        let info = self.local_node.local_chain_info(chain_id).await?;
+        if target_next_block_height <= info.next_block_height {
+            Ok(info)
+        } else {
+            Err(ChainClientError::CannotDownloadCertificates {
+                chain_id,
+                target_next_block_height,
+            })
+        }
     }
 
-    async fn try_process_certificates(
+    /// Downloads and processes all certificates up to (excluding) the specified height from the
+    /// given validator.
+    #[instrument(level = "trace", skip_all)]
+    async fn try_download_certificates_from(
         &self,
-        remote_node: &RemoteNode<P::Node>,
+        remote_node: &RemoteNode<impl ValidatorNode>,
+        chain_id: ChainId,
+        mut start: BlockHeight,
+        stop: BlockHeight,
+    ) -> Result<(), ChainClientError> {
+        while start < stop {
+            // TODO(#2045): Analyze network errors instead of guessing the batch size.
+            let limit = u64::from(stop)
+                .checked_sub(u64::from(start))
+                .ok_or(ArithmeticError::Overflow)?
+                .min(1000);
+            let Some(certificates) = remote_node
+                .try_query_certificates_from(chain_id, start, limit)
+                .await?
+            else {
+                break;
+            };
+            let Some(info) = self
+                .try_process_certificates(remote_node, chain_id, certificates)
+                .await
+            else {
+                break;
+            };
+            assert!(info.next_block_height > start);
+            start = info.next_block_height;
+        }
+        Ok(())
+    }
+
+    #[instrument(level = "trace", skip_all)]
+    pub async fn try_process_certificates(
+        &self,
+        remote_node: &RemoteNode<impl ValidatorNode>,
         chain_id: ChainId,
         certificates: Vec<Certificate>,
     ) -> Option<Box<ChainInfo>> {
-        self.local_node
-            .try_process_certificates(remote_node, chain_id, certificates, &self.notifier)
-            .await
+        let mut info = None;
+        for certificate in certificates {
+            let hash = certificate.hash();
+            if !certificate.value().is_confirmed() || certificate.value().chain_id() != chain_id {
+                // The certificate is not as expected. Give up.
+                warn!("Failed to process network certificate {}", hash);
+                return info;
+            }
+            let mut result = self.handle_certificate(certificate.clone(), vec![]).await;
+
+            result = match &result {
+                Err(err) => {
+                    if let Some(blob_ids) = err.get_blobs_not_found() {
+                        let blobs = remote_node.try_download_blobs(blob_ids.as_slice()).await;
+                        if blobs.len() != blob_ids.len() {
+                            result
+                        } else {
+                            self.handle_certificate(certificate, blobs).await
+                        }
+                    } else {
+                        result
+                    }
+                }
+                _ => result,
+            };
+
+            match result {
+                Ok(response) => info = Some(response.info),
+                Err(error) => {
+                    // The certificate is not as expected. Give up.
+                    warn!("Failed to process network certificate {}: {}", hash, error);
+                    return info;
+                }
+            };
+        }
+        // Done with all certificates.
+        info
     }
 
     async fn handle_certificate(
@@ -498,6 +592,15 @@ pub enum ChainClientError {
 
     #[error("Blobs not found: {0:?}")]
     BlobsNotFound(Vec<BlobId>),
+
+    #[error(
+        "Failed to download certificates and update local node to the next height \
+         {target_next_block_height} of chain {chain_id:?}"
+    )]
+    CannotDownloadCertificates {
+        chain_id: ChainId,
+        target_next_block_height: BlockHeight,
+    },
 }
 
 impl From<Infallible> for ChainClientError {
@@ -903,7 +1006,7 @@ where
     // Verifies that our local storage contains enough history compared to the
     // expected block height. Otherwise, downloads the missing history from the
     // network.
-    async fn synchronize_until(
+    pub async fn synchronize_until(
         &self,
         next_block_height: BlockHeight,
     ) -> Result<Box<ChainInfo>, ChainClientError> {
